Task 0

+++ 0x15-file_io/0-read_textfile.c
@@ -23,7 +23,7 @@ ssize_t read_textfile(const char *filename, size_t letters)
 	if (rd < 0)
 		return (0);
 	wr = write(STDOUT_FILENO, buf, rd);
-	if (wr < 0 || wr != letters)
+	if (wr < 0 || wr != (ssize_t)letters)
 		return (0);
 	close(op);
 	_putchar('\n');
